exercises/exercise4-cicd-security-gates/requirements.txt
Flask==2.0.1
bcrypt==3.2.0
pyjwt==2.0.1